pyproject.toml
[project]
name = "oc-projet-2"
version = "0.1.0"
description = "oc projet 2"
authors = [
    {name = "Pierre Pltn",email = "pierre.pluton@outlook.fr"}
]
package-mode = false
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=2.3.1,<3.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "joblib (>=1.5.1,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
